:lipstick: Renew profile page
